annotations and tests

+++ src/main/java/de/telran/springdemo/service/GreetingService.java
@@ -3,7 +3,7 @@ package de.telran.springdemo.service;
 import de.telran.springdemo.model.Greeting;
 
 public interface GreetingService {
-    int create( Greeting greeting);
+    long create( Greeting greeting);
 
     Greeting get(long id);
